FEATURE add auth hook
