+++ src/features/auth/ui/sign-in/LoginForm.tsx
@@ -1,9 +1,6 @@
 import {yupResolver} from '@hookform/resolvers/yup';
-import {signInWithEmailAndPassword} from 'firebase/auth';
-import {useState} from 'react';
 import {SubmitHandler, useForm} from 'react-hook-form';
-import {useNavigate} from 'react-router-dom';
-import {auth} from 'shared/lib/firebase';
+import {useFirebaseAuth} from 'shared/api/firebase';
 import {Button} from 'shared/ui/button';
 import {FormField} from 'shared/ui/form-field';
 import * as Yup from 'yup';
@@ -24,38 +21,15 @@ const schema = Yup.object({
 });
 
 export const LoginForm = () => {
+    const {loading, loginUser, getErrorMessage} = useFirebaseAuth();
     const {
         register,
         formState: {errors},
         handleSubmit,
     } = useForm<LoginFormValues>({resolver: yupResolver(schema)});
-    const navigate = useNavigate();
-    const [error, setError] = useState('');
-    const [loading, setLoading] = useState(false);
 
     const onSubmit: SubmitHandler<LoginFormValues> = ({email, password}) => {
-        setLoading(true);
-        signInWithEmailAndPassword(auth, email, password).then(user => {
-            // TODO реализовать логику сохранения пользователя
-            console.log(user);
-            navigate('/', {replace: true});
-        }).catch(error => {
-            setError(error.code);
-            setLoading(false);
-        });
-    };
-
-    const showRegistrationError = () => {
-        if (!error)
-            return null;
-        switch (error) {
-            case 'auth/user-not-found':
-                return 'Пользователя с таким Email не существует';
-            case 'auth/wrong-password':
-                return 'Неверный пароль';
-            default:
-                return `Неизвестная ошибка ${error}`;
-        }
+        loginUser(email, password);
     };
 
     return (
@@ -63,7 +37,7 @@ export const LoginForm = () => {
             <div className={styles.title}>
                 <h2>Вход в мой бестиарий</h2>
             </div>
-            <div className={styles.error}>{showRegistrationError()}</div>
+            <div className={styles.error}>{getErrorMessage()}</div>
             <form className={styles.form} onSubmit={handleSubmit(onSubmit)}>
                 <div className={styles.fields}>
                     <FormField label="Email" id="email" type="text" register={register} error={errors.email}/>
