+++ src/features/auth/ui/sign-up/RegistrationForm.tsx
@@ -1,9 +1,6 @@
 import {yupResolver} from '@hookform/resolvers/yup';
-import {createUserWithEmailAndPassword} from 'firebase/auth';
-import {useState} from 'react';
 import {SubmitHandler, useForm} from 'react-hook-form';
-import {useNavigate} from 'react-router-dom';
-import {auth} from 'shared/lib/firebase';
+import {useFirebaseAuth} from 'shared/api/firebase';
 import {Button} from 'shared/ui/button';
 import {FormField} from 'shared/ui/form-field';
 import * as Yup from 'yup';
@@ -33,31 +30,10 @@ export const RegistrationForm = () => {
         handleSubmit,
         formState: {errors},
     } = useForm<RegistrationFormValues>({resolver: yupResolver(schema)});
-    const navigate = useNavigate();
-    const [error, setError] = useState('');
-    const [loading, setLoading] = useState(false);
+    const {loading, getErrorMessage, createNewUser} = useFirebaseAuth();
 
     const onSubmit: SubmitHandler<RegistrationFormValues> = ({email, password}) => {
-        setLoading(true);
-        createUserWithEmailAndPassword(auth, email, password).then(user => {
-            // TODO реализовать логику сохранения пользователя
-            console.log(user);
-            navigate('/', {replace: true});
-        }).catch(error => {
-            setError(error.code);
-            setLoading(false);
-        });
-    };
-
-    const showRegistrationError = () => {
-        if (!error)
-            return null;
-        switch (error) {
-            case 'auth/email-already-in-use':
-                return 'Пользователь с таким Email уже существует';
-            default:
-                return 'Ошибка сервера, попробуйте позже';
-        }
+        createNewUser(email, password);
     };
 
     return (
@@ -65,7 +41,7 @@ export const RegistrationForm = () => {
             <div className={styles.title}>
                 <h2>Регистрация бестиария</h2>
             </div>
-            <div className={styles.error}>{showRegistrationError()}</div>
+            <div className={styles.error}>{getErrorMessage()}</div>
             <form className={styles.form} onSubmit={handleSubmit(onSubmit)}>
                 <div className={styles.fields}>
                     <FormField label="Email" id="email" type="text" register={register} error={errors.email}/>
